refactor: LottoDto dto 패키지로 이동
