+++ src/main/java/lotto/view/InputView.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
-import lotto.model.LottoDto;
+import lotto.dto.LottoDto;
 
 public class InputView {
